chore: separate normalization logic
